+++ lib/stdio/vprintf.c
@@ -2,7 +2,7 @@
 #include <unistd.h>
 #include <printf_common.h>
 
-void vprintf_output(Output op, int *chars_printed, size_t size, char character)
+void vprintf_output(Output op, size_t *chars_printed, size_t size, char character)
 {
     // We always print to stdout: Ignore all limitations
     write(1, &character, 1);
